Cnverted YouTube and Google Documents to new config
Finished the YouTube and Google Documents to the new config
system.
Added disable_check and disabled_value to ConfigItems.
It allows a config item to be enabled and disabled by
a checkbox in the config's label, while returning a specific
value for when it is disabled. It is used in the YouTube
configuration.